// hr-frontend/src/lib/api/generated/model/unitEmployee.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * OpenAPI definition
 * OpenAPI spec version: v0
 */

export interface UnitEmployee {
  id?: number;
  empCode?: string;
  name?: string;
  email?: string;
  phone?: string;
  department?: string;
  position?: string;
  status?: string;
  joinDate?: string;
}